// spring-ai-alibaba-admin/Dify-DSL-Import-Fidelity-Patch/files/frontend/packages/main/src/types/appManage.ts
export interface IGlobalVariableItem {
  key: string;
  type: IValueType;
  desc: string;
  default_value: string | number | boolean | object | null;
}